Everywhere: Add remote desktop support
This adds a RemoteDesktopClient, RemoteDesktopServer, as well as the
LibRemoteDesktop library.

The RemoteDesktop IPC multiplexes different channels, such as RemoteGfx
and RemoteCompositor into one channel so that it can be transported
over one single TCP connection. This is done by simply forwarding the
raw IPC messages wrapped into an appropriate RemoteDesktop IPC message.

When a remote desktop connection is made to RemoteDesktopServer it
will reach out to WindowServer and ask it to start a remote session.
If there is no other remote session in progress, that request is
granted. WindowServer will then notify all clients that a remote
session just started, which causes them to establish a RemoteGfx
connection to RemoteDesktopServer. This enables LibGfx to relay bitmap
painting operations over IPC to RemoteDesktopServer, who then simply
multiplexes them over the same RemoteDesktop IPC connection. When
the session is dropped, the reverse happens and all clients cease to
talk to RemoteDesktopServer.

WindowServer itself for the most part provides Compositor related
information over the RemoteCompositor IPC messages, which consists
largely of window layout, what backing bitmap each window uses, and
input handling. Like regular applications, it also has a RemoteGfx
connection, which are used for rendering window frames, menus and other
WindowServer-managed bitmaps.

On the client side, the RemoteDesktopClient process receives all these
multiplexed IPC messages, which it separates and tracks for each client
application. The RemoteGfx IPC calls mirror each applications actions
on Gfx::Bitmap objects that are relevant for display, and it keeps them
in sync with the application on the remote machine. If needed by the
client, Gfx::Bitmap objects, fonts, etc are transmitted over IPC calls
and the client keeps them around for as long as they are needed.

Then, to render the actual desktop, the RemoteCompositor IPC data is
used to determine where to draw each backing bitmap, window frame
bitmap, etc.