Kernel: Remove g_scheduler_lock
This solves some deadlocks due to multiple locks having to be acquired
at the same time. Also, this allows the system to use a lot more of its
resources because for most things only a thread's m_lock needs to be
acquired. E.g. all processors can switch contexts simultaneously as
they only need to hold the previous and next thread's lock.